Secd 449/db scripts in service (#9)
* Add script runner to the pgsql DB implementation

This adds the ability to execute a SQL script from a file in the repo.
We are using packr to access files which enables us to pack the scripts
in to the final go binary that we deploy rather than needing a
filesystem.

* Remove runhttp and logevent from tests

These are unit tests of components that don't use these dependencies.

* Remove logevent and runhttp from integration tests

These tests do not use these components.

* Add integration tests to linter

* Refactor integration tests to use script runner

Previously, these tests engaged with packr directly to load scripts and
used the sql connection directly to execute them. Now they use the
ScriptRunner addition to the storage.DB.

* Make create script optional and run at init

Add IF NOT EXISTS to DDL statements in 2_create.sql in order to allow
for the script to be run on startup by all nodes.
Add the 2_create.sql script to startup for all nodes.

* Add packr to build process

The packr binary generates all the go files needed in our project based
on inspecting the code for usage. The result is a binary packed version
of our files so that we don't need them on disk.

* Fix test missed in during rebase

After a rebase, one test was still using the unnamed parameter style for
creating a struct instance. A new field was added to the struct which
broke this. This patch converts it to named parameters.

* Fix linting errors after rebase

Some of the changes introduced before this PR contained linter errors.